requirements.txt
django==4.0.5
djangorestframework==3.13.1
django-filter==22.1
psycopg2==2.9.3
#PyJWT==2.4.0
#djangorestframework-simplejwt==5.2.0 # ?????
